old ASPJ working version/ASPJ V41 Final Presentation V1/ASPJ Final/ASPJ V41 Final Presentation V1/requirements.txt
colorlog==6.8.2
cryptography==42.0.8
Flask==3.0.3
Flask_Bcrypt==1.0.1
Flask_Cors==4.0.1
flask_ipban==1.1.5
Flask_Limiter==3.8.0
Flask_Login==0.6.3
Flask_Mail==0.9.1
Flask_MySQLdb==2.0.0
flask_session==0.8.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
itsdangerous==2.1.2
MySQL-python==1.2.5
mysql_connector_repackaged==0.3.1
nest_asyncio==1.6.0
Pillow==10.4.0
plotly==5.18.0
pycryptodome==3.20.0
PyMySQL==1.1.1
pyotp==2.9.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
Requests==2.32.3
stripe==7.12.0
twilio==9.2.3
virustotal_python==1.1.0
vt_py==0.18.2
waitress==3.0.0
Werkzeug==3.0.3
WTForms==3.1.2
